kma/requirements.txt
attrs==25.3.0
beautifulsoup4==4.12.2
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
exceptiongroup==1.2.2
fastapi==0.104.1
h11==0.14.0
idna==3.10
lxml==4.9.3
numpy==1.26.4
outcome==1.3.0.post0
packaging==25.0
pandas==2.1.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.1.0
pytz==2025.1
requests==2.31.0
selenium==4.29.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.7
trio==0.29.0
trio-websocket==0.12.2
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.3.0
uvicorn[standard]==0.24.0
webdriver-manager==4.0.2
websocket-client==1.8.0
wsproto==1.2.0
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1
